+++ client/src/components/index.ts
@@ -5,3 +5,4 @@ export * from './signup';
 export * from './compose';
 export * from './approve';
 export * from './dashboard';
+export * from './certificateView';
